% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{build_doors}
\alias{build_doors}
\title{Changing Number of Doors, Goats and Cars}
\usage{
build_doors(n = 10)
}
\arguments{
\item{...}{no arguments are used by the function.
#' @return}
}
\description{
build_doors will be used to be able to change the number of doors.
}
\details{
We can alter our code to change how many doors, goats, and cars we use. You will #'need to download magrittr package: library(magrittr). For this example we will have 10 doors with 1 cars and 9 goats.
}
