% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.doors, a.pick)

change_door(stay = T, opened.doors, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
return(final.pick)

This returns whether the contestant won or lost
}
\description{
The function change_door is used to create the outcomes for if a contestant stays or #' change the door their initially chose

The code for changing doors and determining a winner are the same as in our first #'example.
}
\details{
The contestant is given the option to change from their initial selection to the other #'door that is still closed.
}
