% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{determine_winner}
\alias{determine_winner}
\title{Determining if Contestant Wins}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a Win or Lose
}
\description{
return(game.results) will determine if the contestant has won or #'  #' lost.
}
\details{
This function will determine if the contestant wins or loses and will #'  report it.
}
\examples{
return(game.results)

}
