% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{open_doors}
\alias{open_doors}
\title{Opening a Door}
\usage{
open_doors(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This determines what what door is opened
}
\description{
The same code we used when using 3 doors is also used for this step.
}
