% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3
}
\description{
The function open_goat_door will open a door with a goat behind it.
}
\details{
The host will always open a door w/ a goat behind it. But it can't be a door the #'contestant has already selected. So it must be a door that is not a car and not a #'current conestant selection.
}
\examples{
return(opened.door)

}
