% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_game}
\alias{play_game}
\title{Playing the Game}
\usage{
play_game()

play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
return(game.results)

This returns the outcome for each strategy
The commands created above are played out in this code. The options are presented here as switching and staying and the whether the outcome was "win" or "lose".
}
\description{
play_game function will go through the code and play the game

play_game function will go through the code and play the game and show the outcome of switching vs staying
}
\details{
The commands created above are played out in this code. The options are presented here as switching and staying and the whether the outcome was "win" or "lose".

The game will be played and it will show us the outcome of switching vs staying
}
\examples{
play_game()

play_game()

}
