% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_n_games}
\alias{play_n_games}
\title{Adding the Game to a Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function will return a table the number of wins and loses in a table
}
\description{
play_n_games function will allow us to play the game multiple times, in this case 100 #'times and see the results.
}
\details{
If we want to see the results for this game after playing the game many times we can #'add the game to a loop. This will play the game 'n" number of times and report the #'results
}
\examples{
(return(results.df))

}
