% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\docType{data}
\name{results.df}
\alias{results.df}
\title{Adding the Game to a Loop}
\format{
An object of class \code{data.frame} with 20000 rows and 2 columns.
}
\usage{
results.df
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function will return a table the number of wins and loses in a table
}
\description{
play_n_games function will allow us to play the game multiple times, in this case 100 #'times and see the results.
}
\details{
If we want to see the results for this game after playing the game many times we can #'add the game to a loop. This will play the game 'n" number of times and report the #'results
}
\examples{
(return(results.df))

}
\keyword{datasets}
