% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{select_door}
\alias{select_door}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{Creating Doors 
}\if{html}{\out{</div>}}}
\usage{
select_door()

select_door()
}
\arguments{
\item{...}{no arguments are used by the function.
A door is picked}
}
\value{
The function returns a number between 1 and 3, that will be the door that is selected.
}
\description{
Using '<-' we will create 3 doors and assign it doors

This function will select a door
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Three doors are created here. The contestant will select a door. 
}\if{html}{\out{</div>}}
}
\examples{
 return(a.pick)
}
