% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\docType{data}
\name{this.game}
\alias{this.game}
\title{Play the Game}
\format{
An object of class \code{character} of length 10.
}
\usage{
this.game
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This code plays the game
}
\description{
To play the game, it is the same code as our first example.
}
\keyword{datasets}
